Updated the unique ID for the entity returned for successful POST requests
Since there was a lot of use of `maxId +1`, I reduced it to a 1-liner to be used throughout.  This was also useful since upon successful creation of the note, the entity is returned as a successful response.

The `nextId` variable was then used to update the entity ID before it is sent in order to notify the user with the updated note ID.