Added new link for auto-focus on input fields
Found an interesting link regarding the use of auto-focus of input
fields using React.createRef().  These refs give us the ability to
reference the DOM element by creating a reference bound to the instance
of the class component.  This could also be applied to functional
components by placing the input field in a functional component and
instantiating that functional component within the class component
instead.